server/requirements.txt
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dlib @ git+https://github.com/davisking/dlib.git@dd1ec1fcff9479b00b233755aa673c028004d96d
face-recognition==1.3.0
face-recognition-models==0.3.0
ffmpeg==1.4
ffmpeg-python==0.2.0
Flask==2.2.3
Flask-Cors==3.0.10
fonttools==4.38.0
future==0.18.2
imageio==2.22.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.0
networkx==2.8.6
numpy==1.24.2
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
scikit-image==0.19.3
scipy==1.9.1
six==1.16.0
tifffile==2022.8.12
Werkzeug==2.2.3
zipp==3.13.0